refactor(map): extract lens row + 3 map overlays (7/7)
Step 7 — the final step of the PulseMapView strangler decomposition.
Pure move; render output is byte-equivalent.

sub/MapLensRow.tsx
  - Lens triad (TODAY/WEEK/ATLAS) + base-tile picker (Map/Sat/Terr/Hybrid)
    extracted as a single self-contained row component. Imports LENS_OPTIONS
    and MAP_VIEW_OPTIONS from ./mapLens directly so the host no longer
    needs them.

overlays/AtlasHalos.tsx
  - Atlas-only density layer. Memoises the halo list off the contacts
    array (stronger radius + opacity for Pulse-linked + team members).
    Renders one Circle per pinned point.

overlays/AtlasTerritories.tsx
  - Atlas-only convex-hull Polygons, one per ContactCircle whose members
    contribute at least 3 pinned points. Click toggles a focused state
    via the supplied onSelectCircle callback.

overlays/AcceptedRoutePolyline.tsx
  - Coral Polyline for an accepted route. Heavy stroke + repeating
    direction dots every 14px so long routes feel oriented at a glance.
    Click handler relays back to the parent's open-in-system-maps.

PulseMapView.tsx
  - Drops the inline atlasDensityHalos memo (moved to AtlasHalos).
  - Replaces the ~95-line inline lens row JSX with <MapLensRow … />.
  - Replaces the inline Polyline / Polygon-loop / Circle-loop blocks
    with <AcceptedRoutePolyline />, <AtlasTerritories />, <AtlasHalos />.
  - Drops Circle, Polygon, Polyline from the @react-google-maps/api
    import, convexHull from mapService, and LENS_OPTIONS + MAP_VIEW_OPTIONS
    from ./sub/mapLens — all only used inside the extracted components now.

After 7 steps: PulseMapView.tsx is 623 lines (down from 2,168 — a 71%
reduction) and the map section is split across 1 host component, 9 hooks,
8 sub-components, and 3 overlays. Named re-exports for the e2e harness
(AiStrip, LiveBroadcastSheet, AiState, AcceptedRoute) are preserved.
